refactor(payload): consolidate all Payload CMS files under lib/payload/
- Move collections/, globals/, migrations/ into lib/payload/
- Move payload.config.ts to lib/payload/ with updated imports
- Move scripts/seed-payload.ts to lib/payload/scripts/
- Add barrel exports for collections and globals
- Configure migrationDir in database adapters for new location
- Update tsconfig @payload-config path alias
- Update package.json seed script path
- Exclude migrations from Biome linting
- Remove obsolete testing-strategy.md doc
