+++ lib/payload/collections/index.ts
@@ -1,0 +1,6 @@
+export { Media } from "./Media";
+export { Projects } from "./Projects";
+export { Roles } from "./Roles";
+export { Testimonials } from "./Testimonials";
+export { Users } from "./Users";
+export { Uses } from "./Uses";
