+++ lib/payload/globals/index.ts
@@ -1,0 +1,1 @@
+export { SiteContent } from "./SiteContent";
