+++ lib/payload/queries/projects.ts
@@ -1,7 +1,7 @@
-import type { Project } from "@/payload-types";
 import { getPayloadClient } from "../get-payload";
+import type { Project } from "../payload-types";
 
-export type { Project as PayloadProject } from "@/payload-types";
+export type { Project as PayloadProject } from "../payload-types";
 
 export async function getAllProjects(): Promise<Project[]> {
 	const payload = await getPayloadClient();
