+++ lib/payload/queries/roles.ts
@@ -1,7 +1,7 @@
-import type { Role } from "@/payload-types";
 import { getPayloadClient } from "../get-payload";
+import type { Role } from "../payload-types";
 
-export type { Role as PayloadRole } from "@/payload-types";
+export type { Role as PayloadRole } from "../payload-types";
 
 export async function getAllRoles(): Promise<Role[]> {
 	const payload = await getPayloadClient();
