+++ lib/payload/queries/testimonials.ts
@@ -1,7 +1,7 @@
-import type { Testimonial } from "@/payload-types";
 import { getPayloadClient } from "../get-payload";
+import type { Testimonial } from "../payload-types";
 
-export type { Testimonial as PayloadTestimonial } from "@/payload-types";
+export type { Testimonial as PayloadTestimonial } from "../payload-types";
 
 export async function getAllTestimonials(): Promise<Testimonial[]> {
 	const payload = await getPayloadClient();
